Move all constant files into a src folder and require
