+++ packages/utilities/psammead-styles/colours.js
@@ -1,15 +1,1 @@
-// BBC News Article colours
-export const C_EBON = '#222222';
-export const C_POSTBOX = '#B80000';
-export const C_STORM = '#404040';
-export const C_WHITE = '#FFFFFF';
-export const C_BLUEJAY = '#0F556C';
-export const C_BLUEJAY_LHT = '#C3DEE7';
-export const C_OAT_LHT = '#F5F3F1';
-export const C_PEBBLE = '#5C5752';
-export const C_RHINO = '#5A5A5A';
-export const C_STONE = '#D5D0CD';
-export const C_CHALK = '#ECEAE7';
-
-// Colours from other BBC services
-export const C_ORBIT_GREY = '#4C4C4C';
+module.exports = require('./dist/colours'); // eslint-disable-line import/no-unresolved
