+++ packages/utilities/psammead-styles/fonts.js
@@ -1,10 +1,1 @@
-// BBC News Article fonts
-const fontFamilyBase = ', Helvetica, Arial, sans-serif';
-export const FF_NEWS_SANS_LIT = `ReithSansNewsLight${fontFamilyBase}`;
-export const FF_NEWS_SANS_REG = `ReithSansNewsRegular${fontFamilyBase}`;
-export const FF_NEWS_SANS_MDM = `ReithSansNewsMedium${fontFamilyBase}`;
-export const FF_NEWS_SANS_BLD = `ReithSansNewsBold${fontFamilyBase}`;
-export const FF_NEWS_SERIF_LIT = `ReithSerifNewsLight${fontFamilyBase}`;
-export const FF_NEWS_SERIF_REG = `ReithSerifNewsRegular${fontFamilyBase}`;
-export const FF_NEWS_SERIF_MDM = `ReithSerifNewsMedium${fontFamilyBase}`;
-export const FF_NEWS_SERIF_BLD = `ReithSerifNewsBold${fontFamilyBase}`;
+module.exports = require('./dist/fonts'); // eslint-disable-line import/no-unresolved
